code refactoring. added traits.
